+++ src/main/java/com/example/finappapirest/finances/interfaces/rest/resources/request/credit/PreviewOnePaymentRequest.java
@@ -4,7 +4,7 @@ import java.time.LocalDate;
 
 public record PreviewOnePaymentRequest(
         Long orderId,
-        Long initialPayment,
+        Float initialPayment,
         CreateInterestRateRequest compensatoryRate,
         CreateInterestRateRequest moratoriumRate,
         LocalDate disbursementDate,
